feat: added ignore button to the mod UI so storages can be set to ignore. Ignored storages will not be considered by the auto-sorters when transfering items.
fix: fixed an issue which transfered items even if the auto sorter inventory was full causing items to get lost on transfer.
fix: fixed a UI issue when interacting with a regular storage immediately after loading the world.
feat: if the storage is picked up you will now get all items back which you used to upgrade the storage to an auto-sorter. This is always 100%, regardless of the "ReturnItemsOnDowngradeMultiplier" setting in the config to make sure storages can easily be moved and upgraded again without loss of resources. On downgrading a auto-sorter to a regular storage the multiplier is still applied.
fix: on downgrade you will now drop all excess items if your inventory is full instead of just losing them.
fix: manual mode is now the default mode for all auto-sorters to make sure you have full control of when the sorter actually starts transferring items. This is especially important if you upgrade a storage which already contains lots of items. This change might set existing auto-sorters, which are in auto mode, to load in manual mode, so make sure you set them back to auto if you want to after loading the world.